+++ lib/cjs/pixelbin.d.ts
@@ -8,40 +8,65 @@ export type transformations = {
      */
     RemoveBG: typeof RemoveBG;
     /**
-     * - Remove background from any image
+     * - EraseBG Background Removal Module
      */
     EraseBG: typeof EraseBG;
     /**
-     * - Upscale image resolution
+     * - Super Resolution Module
      */
     SuperResolution: typeof SuperResolution;
     /**
-     * - Remove JPEG compression artifact noise and get cleaner images
+     * - Artifact Removal Plugin
      */
-    ArtifactRemoval: typeof ArtifactRemoval;
+    Artifact: typeof Artifact;
     /**
-     * - Produce watermark free images
+     * - Watermark Removal Plugin
      */
     WatermarkRemoval: typeof WatermarkRemoval;
     /**
-     * - Detect objects in images.
+     * - Detect objects and text in images
      */
-    AWSRekognition: typeof AWSRekognition;
+    AWSRekognitionPlugin: typeof AWSRekognitionPlugin;
     /**
-     * - Detect objects in images.
+     * - Detect content and text in images
      */
-    GoogleVision: typeof GoogleVision;
+    GoogleVisionPlugin: typeof GoogleVisionPlugin;
+    /**
+     * - Watermark Detection Plugin
+     */
+    WatermarkDetection: typeof WatermarkDetection;
+    /**
+     * - Intelligent Crop Plugin
+     */
+    IntelligentCrop: typeof IntelligentCrop;
+    /**
+     * - OCR Module
+     */
+    TextDetectionandRecognition: typeof TextDetectionandRecognition;
+    /**
+     * - Number Plate Detection Plugin
+     */
+    NumberPlateDetection: typeof NumberPlateDetection;
+    /**
+     * - Image Centering Module
+     */
+    ImageCentering: typeof ImageCentering;
 };
 /**
  * @typedef {Object} transformations
  * @property {Basic} Basic - Basic Transformations
  * @property {RemoveBG} RemoveBG - Remove background from any image
- * @property {EraseBG} EraseBG - Remove background from any image
- * @property {SuperResolution} SuperResolution - Upscale image resolution
- * @property {ArtifactRemoval} ArtifactRemoval - Remove JPEG compression artifact noise and get cleaner images
- * @property {WatermarkRemoval} WatermarkRemoval - Produce watermark free images
- * @property {AWSRekognition} AWSRekognition - Detect objects in images.
- * @property {GoogleVision} GoogleVision - Detect objects in images.
+ * @property {EraseBG} EraseBG - EraseBG Background Removal Module
+ * @property {SuperResolution} SuperResolution - Super Resolution Module
+ * @property {Artifact} Artifact - Artifact Removal Plugin
+ * @property {WatermarkRemoval} WatermarkRemoval - Watermark Removal Plugin
+ * @property {AWSRekognitionPlugin} AWSRekognitionPlugin - Detect objects and text in images
+ * @property {GoogleVisionPlugin} GoogleVisionPlugin - Detect content and text in images
+ * @property {WatermarkDetection} WatermarkDetection - Watermark Detection Plugin
+ * @property {IntelligentCrop} IntelligentCrop - Intelligent Crop Plugin
+ * @property {TextDetectionandRecognition} TextDetectionandRecognition - OCR Module
+ * @property {NumberPlateDetection} NumberPlateDetection - Number Plate Detection Plugin
+ * @property {ImageCentering} ImageCentering - Image Centering Module
  */
 /**
  * class to create a Pixelbin object
@@ -79,10 +104,15 @@ declare class Pixelbin {
         RemoveBG: typeof RemoveBG;
         EraseBG: typeof EraseBG;
         SuperResolution: typeof SuperResolution;
-        ArtifactRemoval: typeof ArtifactRemoval;
+        Artifact: typeof Artifact;
         WatermarkRemoval: typeof WatermarkRemoval;
-        AWSRekognition: typeof AWSRekognition;
-        GoogleVision: typeof GoogleVision;
+        AWSRekognitionPlugin: typeof AWSRekognitionPlugin;
+        GoogleVisionPlugin: typeof GoogleVisionPlugin;
+        WatermarkDetection: typeof WatermarkDetection;
+        IntelligentCrop: typeof IntelligentCrop;
+        TextDetectionandRecognition: typeof TextDetectionandRecognition;
+        NumberPlateDetection: typeof NumberPlateDetection;
+        ImageCentering: typeof ImageCentering;
     };
     static Transformation: typeof Transformation;
     /**
@@ -103,19 +133,29 @@ import * as Basic from "./transformations/Basic";
 import * as RemoveBG from "./transformations/RemoveBG";
 import * as EraseBG from "./transformations/EraseBG";
 import * as SuperResolution from "./transformations/SuperResolution";
-import * as ArtifactRemoval from "./transformations/ArtifactRemoval";
+import * as Artifact from "./transformations/Artifact";
 import * as WatermarkRemoval from "./transformations/WatermarkRemoval";
-import * as AWSRekognition from "./transformations/AWSRekognition";
-import * as GoogleVision from "./transformations/GoogleVision";
+import * as AWSRekognitionPlugin from "./transformations/AWSRekognitionPlugin";
+import * as GoogleVisionPlugin from "./transformations/GoogleVisionPlugin";
+import * as WatermarkDetection from "./transformations/WatermarkDetection";
+import * as IntelligentCrop from "./transformations/IntelligentCrop";
+import * as TextDetectionandRecognition from "./transformations/TextDetectionandRecognition";
+import * as NumberPlateDetection from "./transformations/NumberPlateDetection";
+import * as ImageCentering from "./transformations/ImageCentering";
 export namespace transformations {
     export { Basic };
     export { RemoveBG };
     export { EraseBG };
     export { SuperResolution };
-    export { ArtifactRemoval };
+    export { Artifact };
     export { WatermarkRemoval };
-    export { AWSRekognition };
-    export { GoogleVision };
+    export { AWSRekognitionPlugin };
+    export { GoogleVisionPlugin };
+    export { WatermarkDetection };
+    export { IntelligentCrop };
+    export { TextDetectionandRecognition };
+    export { NumberPlateDetection };
+    export { ImageCentering };
 }
 import Image from "./image.js";
 export { Pixelbin as default, Transformation };