+++ lib/cjs/transformations/EraseBG.d.ts
@@ -1,5 +1,6 @@
 export function bg(config?: {
     industryType: string;
+    addShadow: boolean;
 }): Transformation;
 declare namespace _default {
     export { bg };