+++ lib/esm/transformations/WatermarkRemoval.d.ts
@@ -1,4 +1,12 @@
-export function remove(config?: {}): Transformation;
+export function remove(config?: {
+    removeText: boolean;
+    removeLogo: boolean;
+    box1: string;
+    box2: string;
+    box3: string;
+    box4: string;
+    box5: string;
+}): Transformation;
 declare namespace _default {
     export { remove };
 }