+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 4.0.0
+
+-   Transformations are updated to their latest API
+-   Added new plugins `Watermark Detection`, `TextDetectionandRecognition`, `NumberPlateDetection`, `IntelligentCrop`, `ImageCentering`
+-   `getUrl` now returns CDN version `v2` urls
+
 # 3.0.0
-* Reduce bundled size from 61kB (minified and gzipped) to 14kB (minified and gzipped)
-* Remove direct import of `core-js` and `regenerator-runtime`
+
+-   Reduce bundled size from 61kB (minified and gzipped) to 14kB (minified and gzipped)
+-   Remove direct import of `core-js` and `regenerator-runtime`